Add support for Bing image driver using full cookie string
This commit introduces improved handling for the Bing image driver which now accepts a full cookie string as a method of authorization. This update includes changes in the environment variables, updates to `BingDallE3Driver` class in the `drivers/bing_dall_e3.py` to handle images using all cookies, and updates in the `README.md` and `action.yml` to reflect these changes. The `fake-useragent` dependency was also added to the `requirements.txt`. Please note, usage of individual `BING_AUTH_TOKEN` or `BING_AUTH_TOKEN_KIEV` is now marked as deprecated, recommend to use the full cookie string.